requirements.txt
fastapi~=0.63.0
uvicorn~=0.13.3

python-dotenv~=0.15.0
exfinance~=0.5.0